refactor: split AdminDashboardPage into feature modules
